requirements.txt
google-generativeai

langchain
PyPDF2
chromadb
faiss-cpu
Flask
python-dotenv
PIL